Fixed timezone bug for Australia, India, or anywhere that uses a fraction of an hour offset from Greenwich, like 9.5
git-svn-id: https://sondow.googlecode.com/svn/trunk@73 9cedc2b2-b939-11de-82ee-51df2961d3cf

+++ sondow/src/com/sondow/CheckTimeZoneServlet.java
@@ -26,7 +26,7 @@ public class CheckTimeZoneServlet extends BaseServlet {
             HttpServletResponse response) throws IOException, ServletException {
 
         String timezoneOffsetString = request.getParameter("timeZoneOffset");
-        int timezoneOffset = Integer.parseInt(timezoneOffsetString);
+        double timezoneOffset = Double.parseDouble(timezoneOffsetString);
 
         // Only show the goods to people who are more than 1 time zone away
         // from the east coast.
